refactor: move Toast into separate component
